+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 - 1.0-BETA1 (2015-01-xx)
+  - renamed ApiErrorException to RequestException and changed its constructor signature
   - now uses Money class for dealing with money
   - Exception now extends \RuntimeException
   - now supports GuzzleHttp client as implementation of ClientInterface